Use SSL for the client-proxy connection
This finally makes the first hop secure, protecting the
initial HTTP CONNECT from sniffing or tampering.

We only include a self-signed cert and corresponding keystore,
and have to make the proxy connection with our test utility insecure,
because there's not a clear way to make the cert's hostname match,
even when we tell curl it's a trusted CA cert. An actual deployment
would need a real cert.